Updated libgpg-error and recompiled dependent packages. 	new file:   .gitignore 	modified:   ts/bin/release 	modified:   ts/build/boot-images/templates/syslinux/installer/menu/common.inc 	modified:   ts/build/build 	modified:   ts/build/conf/devstation/build.conf.example 	modified:   ts/build/conf/installer/build.conf.example 	modified:   ts/build/packages/geany/build/extra/etc/skel/.config/geany/geany.conf 	modified:   ts/build/packages/geany/build/install 	modified:   ts/build/packages/git/build/extra/bin/proxy-setup 	modified:   ts/build/packages/installer/bin/install.sh 	modified:   ts/build/packages/installer/install/syslinux.tpl 	modified:   ts/ports/chroot/libmicrohttpd/.footprint 	modified:   ts/ports/chroot/libmicrohttpd/.md5sum 	modified:   ts/ports/chroot/libmicrohttpd/Pkgfile 	new file:   ts/ports/chroot/libmicrohttpd/libmicrohttpd#0.9.39-1.pkg.tar.gz 	deleted:    ts/ports/chroot/libmicrohttpd/libmicrohttpd#0.9.9-1.pkg.tar.gz 	new file:   ts/ports/components/cryptodev-linux/.footprint 	new file:   ts/ports/components/cryptodev-linux/.md5sum 	new file:   ts/ports/components/cryptodev-linux/Pkgfile 	new file:   ts/ports/components/cryptodev-linux/cryptodev-linux#1.7-3.14.32.pkg.tar.gz 	modified:   ts/ports/components/cryptsetup/cryptsetup#1.5.0-1.pkg.tar.gz 	modified:   ts/ports/components/tigervnc/tigervnc#1.1.0-1.pkg.tar.gz 	modified:   ts/ports/components/wireshark/.footprint 	modified:   ts/ports/components/wireshark/.md5sum 	modified:   ts/ports/components/wireshark/Pkgfile 	renamed:    ts/ports/components/wireshark/wireshark#1.8.4-1.pkg.tar.gz -> ts/ports/components/wireshark/wireshark#1.12.3-1.pkg.tar.gz 	modified:   ts/ports/extensions/geany-plugins/Pkgfile 	modified:   ts/ports/extensions/geany-plugins/geany-plugins#git-1.pkg.tar.gz 	modified:   ts/ports/extensions/gpgme/gpgme#1.3.2-1.pkg.tar.gz 	modified:   ts/ports/extensions/libimobiledevice/libimobiledevice#1.0.7-1.pkg.tar.gz 	modified:   ts/ports/extensions/rasqal/rasqal#0.9.30-1.pkg.tar.gz 	modified:   ts/ports/gnome/evince/evince#2.32.0-1.pkg.tar.gz 	modified:   ts/ports/gnome/glib-networking/glib-networking#2.32.3-1.pkg.tar.gz 	modified:   ts/ports/gnome/gnome-disk-utility/gnome-disk-utility#2.30.1-1.pkg.tar.gz 	modified:   ts/ports/gnome/gnome-keyring/gnome-keyring#2.31.4-1.pkg.tar.gz 	modified:   ts/ports/gnome/gnome-python/gnome-python#2.28.1-1.pkg.tar.gz 	modified:   ts/ports/gnome/gvfs/gvfs#1.12.3-1.pkg.tar.gz 	modified:   ts/ports/gnome/libgnome-keyring/libgnome-keyring#2.31.92-1.pkg.tar.gz 	modified:   ts/ports/gnome/libgnomeui/libgnomeui#2.24.5-1.pkg.tar.gz 	modified:   ts/ports/gnome/libsecret/libsecret#0.11-1.pkg.tar.gz 	modified:   ts/ports/gnome/libsoup/libsoup#2.38.1-1.pkg.tar.gz 	modified:   ts/ports/gnome/networkmanager-applet/networkmanager-applet#0.9.6.4-1.pkg.tar.gz 	modified:   ts/ports/gnome/totem-pl-parser/totem-pl-parser#2.32.6-1.pkg.tar.gz 	modified:   ts/ports/gnome/totem/totem#2.32.0-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/VERSION 	modified:   ts/ports/kernel-modules/aufs/.footprint 	modified:   ts/ports/kernel-modules/aufs/.md5sum 	modified:   ts/ports/kernel-modules/aufs/aufs#3.14.21+-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/bbswitch/.footprint 	modified:   ts/ports/kernel-modules/bbswitch/bbswitch#0.8-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/kernel-TS/.footprint 	modified:   ts/ports/kernel-modules/kernel-TS/.md5sum 	modified:   ts/ports/kernel-modules/kernel-TS/Pkgfile 	modified:   ts/ports/kernel-modules/kernel-TS/kernel-TS#1-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/kernel-TS/ts.config 	modified:   ts/ports/kernel-modules/kernel-TS_64/.footprint 	modified:   ts/ports/kernel-modules/kernel-TS_64/.md5sum 	modified:   ts/ports/kernel-modules/kernel-TS_64/kernel-TS_64#1-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/kernel-TS_64/ts.config 	modified:   ts/ports/kernel-modules/kernel-TS_SMP/.footprint 	modified:   ts/ports/kernel-modules/kernel-TS_SMP/.md5sum 	modified:   ts/ports/kernel-modules/kernel-TS_SMP/kernel-TS_SMP#1-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/kernel-TS_SMP/ts.config 	modified:   ts/ports/kernel-modules/open-vm-tools/.footprint 	modified:   ts/ports/kernel-modules/open-vm-tools/open-vm-tools#9.4.6-1770165-1.pkg.tar.gz 	modified:   ts/ports/kernel-modules/usb-redirector/.footprint 	modified:   ts/ports/kernel-modules/usb-redirector/usb-redirector#3.1-1.pkg.tar.gz 	modified:   ts/ports/proprietary/broadcom-sta/.footprint 	modified:   ts/ports/proprietary/broadcom-sta_64/.footprint 	modified:   ts/ports/xfce/xfwm4-themes/xfwm4-themes#4.10.0-1.pkg.tar.gz 	modified:   usr/ports/opt/git/git#1.9.1-1.pkg.tar.gz 	new file:   usr/ports/opt/gnupg/.footprint 	new file:   usr/ports/opt/gnupg/.md5sum 	new file:   usr/ports/opt/gnupg/Pkgfile 	new file:   usr/ports/opt/gnupg/README 	new file:   usr/ports/opt/gnupg/gnupg#2.0.26-1.pkg.tar.gz 	modified:   usr/ports/opt/gnutls/Pkgfile 	modified:   usr/ports/opt/gnutls/gnutls#2.12.21-1.pkg.tar.gz 	modified:   usr/ports/opt/libassuan/libassuan#2.1.3-1.pkg.tar.gz 	modified:   usr/ports/opt/libgcrypt/libgcrypt#1.5.0-1.pkg.tar.gz 	modified:   usr/ports/opt/libgpg-error/.footprint 	modified:   usr/ports/opt/libgpg-error/.md5sum 	modified:   usr/ports/opt/libgpg-error/Pkgfile 	deleted:    usr/ports/opt/libgpg-error/libgpg-error#1.10-1.pkg.tar.gz 	new file:   usr/ports/opt/libgpg-error/libgpg-error#1.17-3.pkg.tar.gz 	new file:   usr/ports/opt/libksba/.footprint 	new file:   usr/ports/opt/libksba/.md5sum 	new file:   usr/ports/opt/libksba/Pkgfile 	new file:   usr/ports/opt/libksba/libksba#1.3.2-1.pkg.tar.gz 	new file:   usr/ports/opt/libpth/.footprint 	new file:   usr/ports/opt/libpth/.md5sum 	new file:   usr/ports/opt/libpth/Pkgfile 	new file:   usr/ports/opt/libpth/libpth#2.0.7-1.pkg.tar.gz 	modified:   usr/ports/opt/libxslt/libxslt#1.1.26-1.pkg.tar.gz

+++ ts/build/packages/installer/bin/install.sh
@@ -108,7 +108,9 @@ if [ -e /mnt/cdrom0/thindev-default-$TS_VERSION.tar.xz ]; then
 	tar -xvf /mnt/cdrom0/thindev-default-$TS_VERSION.tar.xz
 else
 	echo "Downloading a Default Image"
-	wget "$WEBUPDATEROOT/thindev-default-$TS_VERSION.tar.xz"
+	if ! wget -t 3 -T 30 "$WEBUPDATEROOT/thindev-default-$TS_VERSION.tar.xz"; then
+		exit 2
+	fi
 	tar -xvf thindev-default-$TS_VERSION.tar.xz
 	rm thindev-default-$TS_VERSION.tar.xz
 fi
